requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
mysqlclient
drf-yasg
